requirements.txt
websockets
aiohttp
asyncio
requests
maim_message
loguru
pillow
tomlkit
rich
sqlmodel
grpcio
grpcio-tools
protobuf